Revert "Merge branch 'develop' onto 'master'"
This reverts commit b342b06c5f7540fb0fe5fc3d859aa66f35a12004, reversing
changes made to cb7dde136658b7494381c3a59145ea73bdb48551.

+++ willow/src/transforms/transformbuilder.cpp
@@ -558,7 +558,7 @@ TransformBuilder::split(TensorId in,
   std::vector<TensorId> inputs = {in};
   return multiOutputOp(Onnx::Operators::Split_11,
                        inputs,
-                       static_cast<int>(splitSizes.size()),
+                       splitSizes.size(),
                        {{"axis", axis}, {"split", splitSizes}},
                        virtualGraphId,
                        pipelineStage,
